# fix to_scalar issue of very high latency