I'm hoping to resolve this issue.
Since the dimension of the tensor you wish to call to_scalar is 0. 
So I changed the match arm of `CudaStorage` in `to_scalar` by directly calling `dtoh_sync_copy` on a `CudaView<T>` created designated for copy to CPU mem.
Calling `dtoh_sync_copy` directly with `CudaSlice<T>` will cause a full copy on the underlying storage. By calling CudaView<T> will resolve this problem.

Code sample.

```rust
...
CudaStorageSlice::F32(slice) => {
    let sub_slice = slice.slice(offset..offset + 1);
    let device = slice.device();
    let cpu_storage = device.dtoh_sync_copy(&sub_slice).w()?;
    return from_cpu_storage_single(&CpuStorage::F32(cpu_storage));
}
...
```

Fix #2239