@RoggeOhta 2024-06-03
> @RoggeOhta do you measure a performance increase when you run with `CUDA_LAUNCH_BLOCKING=1`?

Output of Minimal reproducible example in the linked issue:
All build with release profile.
It's a simple test without serious benchmarking, it can only prove the change in the order of magnitudes.

Before fix with CUDA_LAUNCH_BLOCKING=1
```
Tensor on CPU: Tensor[dims 10000, 10000; f32]
117.986679ms
Tensor on GPU: Layout { shape: [10000, 10000], stride: [10000, 1], start_offset: 0 }
4.219µs
259.142736ms
[candle-test/src/main.rs:26:5] scalar = 0.09058428
```

Before fix without CUDA_LAUNCH_BLOCKING=1
```
Tensor on CPU: Tensor[dims 10000, 10000; f32]
92.308598ms
Tensor on GPU: Layout { shape: [10000, 10000], stride: [10000, 1], start_offset: 0 }
7.003µs
292.98987ms
[candle-test/src/main.rs:26:5] scalar = 0.630551
```

After fix with CUDA_LAUNCH_BLOCKING=1
```
Tensor on CPU: Tensor[dims 10000, 10000; f32]
92.458806ms
Tensor on GPU: Layout { shape: [10000, 10000], stride: [10000, 1], start_offset: 0 }
2.835µs
37.574µs # to_scalar
[candle-test/src/main.rs:26:5] scalar = 0.23715067
```

After fix without CUDA_LAUNCH_BLOCKING=1
```
Tensor on CPU: Tensor[dims 10000, 10000; f32]
159.794325ms
Tensor on GPU: Layout { shape: [10000, 10000], stride: [10000, 1], start_offset: 0 }
10.94µs # to_scalar
24.328µs 
[candle-test/src/main.rs:26:5] scalar = 0.076687455
```

Before fix with CUDA_LAUNCH_BLOCKING=1